benchmark/requirements.txt
claude-agent-sdk>=0.1.0
python-dotenv>=1.0.0
httpx>=0.28.0
